# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentFlow API
  version: 0.1.0
  description: |
    Authoritative HTTP contract shared by generated Go DTOs and the typed
    TypeScript client. Business handlers remain implemented with net/http.
    With AUTH_MODE=oidc, every business operation requires a revocable session
    cookie and owned/granted Workspace entity (401/404 on failure). Mutations also
    require the configured frontend Origin. Health/authentication and preflight
    are exempt. Local mode is unauthenticated trusted development, not an
    alternate authentication mechanism. HTTPS uses __Host-agentflow_session;
    loopback HTTP uses agentflow_session. Archived business writes return 409;
    missing scope resolves to the owner's active default. Workspace management
    is owner-scoped independently of the selected header. Agent/Tool configuration
    writes remain trusted-local operator operations (OIDC users receive 403).
    Resource IDs are resolved inside that authorized Workspace; foreign IDs return
    404. SSE delivery rechecks the session and Workspace on each write, including
    Chat/Continue/Resume. A later access failure emits an error frame using the
    existing ChatStreamEvent error contract and stops delivery, without canceling
    the admitted durable Run. See the resource authorization guide.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/workspaces:
    get:
      operationId: listWorkspaces
      tags: [Workspaces]
      description: Lists only the current user's owned, granted, non-deleted Workspaces, including archived read-only spaces. Local mode uses a persistent server-owned local User. Selected Workspace headers do not scope this endpoint.
      responses:
        '200':
          description: Owned Workspaces.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workspace' }
    post:
      operationId: createWorkspace
      tags: [Workspaces]
      description: Owner is derived from the session, never accepted from the client. IDs are generated by PostgreSQL; names do not identify resources.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceCreateRequest' }
      responses:
        '201':
          description: Created Workspace.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/workspaces/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, pattern: '^[1-9][0-9]*$' }
    get:
      operationId: getWorkspace
      tags: [Workspaces]
      responses:
        '200':
          description: Owned and granted Workspace.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateWorkspace
      tags: [Workspaces]
      description: Renames, updates description, archives/restores or selects the default. Closing the current default requires an owned active replacement. Unfinished Runs and uncertain effects block archive/delete.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceUpdateRequest' }
      responses:
        '200':
          description: Updated Workspace, preserving its ID.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Last active space, invalid replacement default, unfinished Run or uncertain effect prevents the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      operationId: deleteWorkspace
      tags: [Workspaces]
      description: Sets deleted_at. Does not physically erase resource data. Deleted spaces cannot be accessed or implicitly recreated at login.
      parameters:
        - in: query
          name: replacement_workspace_id
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
      responses:
        '204': { description: Soft-deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Workspace cannot currently be deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/chat:
    post:
      operationId: streamChat
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Server-sent stream of ChatStreamEvent values.
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/runs/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      operationId: cancelRun
      tags: [Runs]
      responses:
        '200':
          description: Canceled or canceling run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/tools:
    get:
      operationId: listTools
      tags: [Tools]
      description: Service-owned Tool catalog with the selected Workspace's persisted grants and effective availability.
      responses:
        '200':
          description: Tool catalog, Workspace configuration revision and exclusion reasons.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInfo'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tools/{name}/enable:
    parameters:
      - $ref: '#/components/parameters/ToolName'
    post:
      operationId: enableTool
      tags: [Tools]
      description: Allow an installed, service-ready Tool in the selected Workspace. Does not modify operator configuration or Agent selection.
      responses:
        '200':
          description: Updated tool catalog.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tools/{name}/disable:
    parameters:
      - $ref: '#/components/parameters/ToolName'
    post:
      operationId: disableTool
      tags: [Tools]
      description: Revoke the selected Workspace's Tool grant. Subsequent execution and recovery callbacks recheck the grant.
      responses:
        '200':
          description: Updated tool catalog.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
...
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    ToolName:
      name: name
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Run admission queue is full or temporarily unavailable.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Workspace:
      type: object
      required: [id, owner_user_id, name, description, status, is_default, created_at, updated_at]
      properties:
        id: { type: string, pattern: '^[1-9][0-9]*$', description: Decimal BIGINT identity encoded as a string to preserve precision. }
        owner_user_id: { type: string }
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 2000 }
        status: { type: string, enum: [active, archived] }
        is_default: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time }
    WorkspaceCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 2000 }
    WorkspaceUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 2000 }
        status: { type: string, enum: [active, archived] }
        make_default: { type: boolean }
        replacement_workspace_id: { type: string, pattern: '^[1-9][0-9]*$' }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        source:
          type: string
        category:
          type: string
        retryable:
          type: boolean
        operation:
          type: string
        request_id:
          type: string
...
    ChatMode:
      type: string
      enum: [single, multi_agent, autonomous]
    RunStatus:
      type: string
      enum: [queued, running, waiting_for_user, completed, failed, failed_recoverable, canceling, canceled]
    VerificationStatus:
      type: string
      enum: [not_required, pending, running, passed, failed, blocked, stale]
    VerifierType:
      type: string
      enum: [command, http, json_schema, text_constraints, citation, answer_relevance, grounded_answer]
    VerifierSpecInput:
      type: object
      required: [id, type, required, config]
      properties:
        id:
          type: string
          minLength: 1
        type:
          $ref: '#/components/schemas/VerifierType'
        version:
          type: string
        required:
          type: boolean
        timeout_ms:
          type: integer
          format: int64
          minimum: 0
        config:
          type: object
          additionalProperties: true
    VerificationPolicyInput:
      type: object
      required: [mode, max_attempts, on_exhausted]
      properties:
        mode:
          type: string
          enum: [all_must_pass, any_may_pass]
        max_attempts:
          type: integer
          minimum: 1
        on_exhausted:
          type: string
          enum: [fail, waiting_for_user]
    CompletionContractInput:
      type: object
      required: [subject_type, verifiers, policy]
      properties:
        subject_type:
          type: string
          enum: [run_output]
        verifiers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/VerifierSpecInput'
        policy:
          $ref: '#/components/schemas/VerificationPolicyInput'
    CompletionContract:
      allOf:
        - $ref: '#/components/schemas/CompletionContractInput'
        - type: object
          required: [id, version, hash]
          properties:
            id:
              type: string
            version:
              type: integer
            hash:
              type: string
    ChatRequest:
      type: object
      required: [message]
      properties:
        workspace_id:
          type: string
        conversation_id:
          type: string
        agent_id:
          type: string
        message:
          type: string
          minLength: 1
        mode:
          $ref: '#/components/schemas/ChatMode'
        completion_contract:
          $ref: '#/components/schemas/CompletionContractInput'
...
    ToolInfo:
      type: object
      required: [name, description, parameters, enabled]
      properties:
        service_enabled:
          type: boolean
          description: Operator-owned service switch. Workspace owners cannot override it.
        workspace_enabled:
          type: boolean
          description: Persisted Workspace allowlist entry, independent of service readiness.
        config_revision:
          type: integer
          format: int64
        excluded_reason:
          type: string
          description: Service, prerequisite, Workspace or static-policy exclusion reason; argument-dependent decisions remain in Run events.
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        unavailable_reason:
          type: string
          description: Runtime prerequisite that prevents an enabled Tool from being offered to the model.
    Run:
      type: object
      required: [id, workspace_id, agent_id, conversation_id, status, verification_status, active_runtime_ms, created_at, updated_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        agent_id:
          type: string
        conversation_id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        completion_contract:
          $ref: '#/components/schemas/CompletionContract'
        error:
          type: string
        started_at:
          type: string
          format: date-time
        execution_started_at:
          type: string
          format: date-time
        active_runtime_ms:
          type: integer
          format: int64
        heartbeat_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...